+++ Tests/helper.py
@@ -165,12 +165,6 @@ def assert_tuple_approx_equal(actuals, targets, threshold, msg):
     assert value, msg + ": " + repr(actuals) + " != " + repr(targets)
 
 
-def skip_known_bad_test(msg=None):
-    # Skip if PILLOW_RUN_KNOWN_BAD is not true in the environment.
-    if not os.environ.get("PILLOW_RUN_KNOWN_BAD", False):
-        pytest.skip(msg or "Known bad test")
-
-
 def skip_unless_feature(feature):
     reason = "%s not available" % feature
     return pytest.mark.skipif(not features.check(feature), reason=reason)